requirements.txt
eth-brownie
python-dotenv
eth-brownie
